faceDetection/requirements.txt
facenet-pytorch
argparse
numpy
opencv-contrib-python
pillow
